Use fb_fix_screeninfo.line_length rather than virtual resolution width.
With more testing, it looks like using the virtual resolution width
(`fb_var_screeninfo.yres_virtual`) to compute pixel address in the framebuffer
memory buffer was still incorrect, because there could be padding at the
end of each line even if the virtual resolution is equal to the visible
resolution.

In this diff we switch to using `fb_fix_screeninfo.line_length` for
pixel address computation instead.