Fix weird box chars in Linux console with UTF-8 locales.
Prior to this fix, when running in the Linux console with UTF-8 locales
(the default in most modern distros), box characters end up displaying
as "x" and "q" characters.

The fix is based on the following post:

  https://lists.gnu.org/archive/html/bug-ncurses/2015-02/msg00010.html

Instead of linking against libncurses, we need to link against
libncursesw instead. Unfortunately this is only supported in CMake 3.10
and above. Since ultimately the weird chars don't affect usage, I'm not
going to try to work around this limitation for now.