Fix Travis CI deploy to GCS failure on ARM64 due to PR #46.
In PR #46, we switched to using dpl v2 for uploading build artifacts to
Google Cloud Storage. However this results in the following error on
ARM64 platforms:

```
$ curl -L https://dl.google.com/dl/cloudsdk/channels/rapid/google-cloud-sdk.tar.gz | tar xz -C ~ && ~/google-cloud-sdk/install.sh --path-update false --usage-reporting false --command-completion false
Welcome to the Google Cloud SDK!
ERROR: (gcloud.components.update) The following components are unknown [anthoscli].
This will install all the core command line tools necessary for working with
the Google Cloud Platform.
```